Sort accprof table

+++ src/accprof/accprofiling_types.h
@@ -9,7 +9,7 @@ typedef struct {
    int line_end;
    int func_line_start;
    int func_line_end;
-   size_t copied_bytes; // Only for data events: Amount of data moved
+   long long copied_bytes; // Only for data events: Amount of data moved
    char *source_file;
    char *func_name;
    char *var_name; // Only for data events: Name of moved variable
